factor out common test setup code in HelpRankTestHelper
